Move app-level admin template to admin blueprint
